services/dq_report/requirements.txt
kafka-python==2.0.2
fastapi==0.104.1
uvicorn==0.24.0
python-json-logger==2.0.7
